wp: add auth jwt
