+++ requirements.txt
@@ -6,6 +6,7 @@ Django==3.1.4
 django-environ==0.4.5
 django-filter==2.4.0
 djangorestframework==3.12.2
+djangorestframework-camel-case==1.2.0
 gunicorn==20.0.4
 importlib-metadata==3.3.0
 isort==5.6.4
