+++ users/models.py
@@ -1,36 +1,7 @@
 from django.contrib.auth.models import AbstractBaseUser, BaseUserManager, PermissionsMixin
 from django.db import models
-from django.utils import timezone
 
-
-class UserManager(BaseUserManager):
-
-  def _create_user(self, email, password, is_staff, is_superuser, **extra_fields):
-    if not email:
-        raise ValueError('Users must have an email address')
-    now = timezone.now()
-    email = self.normalize_email(email)
-    user = self.model(
-        email=email,
-        is_staff=is_staff, 
-        is_active=True,
-        is_superuser=is_superuser, 
-        last_login=now,
-        date_joined=now, 
-        **extra_fields
-    )
-    user.set_password(password)
-    user.save(using=self._db)
-    return user
-
-  def create_user(self, email, password, **extra_fields):
-    return self._create_user(email, password, False, False, **extra_fields)
-
-  def create_superuser(self, email, password, **extra_fields):
-    user=self._create_user(email, password, True, True, **extra_fields)
-    return user
-
-
+from .managers import UserManager
 class User(AbstractBaseUser, PermissionsMixin):
     email = models.EmailField(max_length=254, unique=True)
     name = models.CharField(max_length=254, null=True, blank=True)
